requirements.txt
mecab-python3==1.0.9
